门店/总店有称重菜时排班参考统计表中已点餐金额和结账金额未按实际称重来计算 10927

+++ newspicyway/sql_script_全量/V100_0_0_1__procedure-report.sql
@@ -7916,7 +7916,7 @@ BEGIN -- 返回字段说明如下
   (
     orderid VARCHAR(50),
     dishid VARCHAR(50),
-    dishnum INT,
+    dishnum DOUBLE(13,2),
     orignalprice DECIMAL(13, 2),
     begintime DATETIME,
     endtime DATETIME
